python_sms_responder/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
twilio==8.1.0
openai==1.7.0
pydantic==1.10.7
python-multipart==0.0.6
psycopg2-binary==2.9.6
